Convert back to doubles.
Now we have the refactoring into the typedefs, but also a somewhat
better speed. The previous commit made matters a little worse.

+++ fc-solve/source/rate_state.h
@@ -38,7 +38,7 @@ extern "C"
 
 #define FCS_BEFS_SEQS_OVER_RENEGADE_CARDS_EXPONENT 1.3
 
-typedef int fc_solve_seq_cards_power_type_t;
+typedef double fc_solve_seq_cards_power_type_t;
 extern fc_solve_seq_cards_power_type_t fc_solve_seqs_over_cards_lookup[2 * 13 * 4 + 1];
 
 #if 0
